backend/api-gateway/api-gateway/requirements.txt
blinker==1.9.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cryptography==36.0.2
Flask==3.1.0
flask-cors==6.0.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pycparser==2.22
PyJWT==2.10.1
requests==2.32.4
SQLAlchemy==2.0.40
typing_extensions==4.14.0
urllib3==2.4.0
Werkzeug==3.1.3
